backend/pyproject.toml
[project]
name = "zara-learns-web"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.9.2",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "tortoise-orm[asyncpg]>=0.25.3",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.aerich]
tortoise_orm = "zara_learns_web.storage.TORTOISE_ORM"
location = "./src/zara_learns_web/migrations"
src_folder = "./."
